requirements.txt
altair==4.2.0
attrs==22.1.0
backports.zoneinfo==0.2.1
blinker==1.5
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
commonmark==0.9.1
confection==0.0.3
cymem==2.0.7
Cython==0.29.32
decorator==5.1.1
entrypoints==0.4
faiss-cpu==1.7.2
filelock==3.8.0
flatbuffers==22.10.26
gitdb==4.0.9
GitPython==3.1.29
huggingface-hub==0.10.1
humanfriendly==10.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
langcodes==3.3.0
lightgbm==3.3.3
MarkupSafe==2.1.1
mpmath==1.2.1
murmurhash==1.0.9
nmslib==2.1.1
numpy==1.21.6
onnxruntime==1.13.1
packaging==21.3
pandas==1.3.5
pathy==0.6.2
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
preshed==3.0.8
protobuf==3.20.3
psutil==5.9.3
pyarrow==10.0.0
pybind11==2.6.1
pydantic==1.10.2
pydeck==0.8.0b4
Pygments==2.13.0
pyjnius==1.4.2
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
pyserini==0.18.0
python-dateutil==2.8.2
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
rich==12.6.0
scikit-learn==1.0.2
scipy==1.7.3
semver==2.13.0
sentencepiece==0.1.97
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.4.2
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
streamlit==1.14.0
sympy==1.10.1
thinc==8.1.5
threadpoolctl==3.1.0
tokenizers==0.13.1
toml==0.10.2
toolz==0.12.0
torch==1.13.0
tornado==6.2
tqdm==4.64.1
transformers==4.23.1
typer==0.4.2
typing_extensions==4.1.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
wasabi==0.10.1
zipp==3.10.0
